Fix image and taxon image serializer specs
