+++ api/spec/serializers/spree/api/v2/platform/image_serializer_spec.rb
@@ -20,6 +20,7 @@ describe Spree::Api::V2::Platform::ImageSerializer do
             alt: image.alt,
             created_at: image.created_at,
             updated_at: image.updated_at,
+            transformed_url: image.generate_url(size: {}),
             original_url: image.original_url
           },
           relationships: {
