+++ core/app/models/spree/image.rb
@@ -1,7 +1,6 @@
 module Spree
-  class Image < Asset
+  class Image < ImageAsset
     include Configuration::ActiveStorage
-    include Rails.application.routes.url_helpers
 
     # In Rails 5.x class constants are being undefined/redefined during the code reloading process
     # in a rails development environment, after which the actual ruby objects stored in those class constants
@@ -49,25 +48,6 @@ module Spree
 
     def plp_url
       generate_url(size: self.class.styles[:plp_and_carousel])
-    end
-
-    def generate_url(size:, gravity: 'center', quality: 80, background: 'show')
-      return if size.blank?
-      size = size.gsub(/\s+/, '')
-
-      return unless size.match(/(\d+)x(\d+)/)
-
-      polymorphic_path(attachment.variant(
-        gravity: gravity,
-        resize: size,
-        extent: size,
-        background: background,
-        quality: quality.to_i
-      ), only_path: true)
-    end
-
-    def original_url
-      polymorphic_path(attachment, only_path: true)
     end
   end
 end
