+++ core/app/models/spree/taxon_image/configuration/active_storage.rb
@@ -1,5 +1,5 @@
 module Spree
-  class TaxonImage < Asset
+  class TaxonImage < ImageAsset
     module Configuration
       module ActiveStorage
         extend ActiveSupport::Concern
